+++ action_test.go
@@ -1392,7 +1392,7 @@ var tests = []struct {
 
 func TestGenerateActions(t *testing.T) {
 	cmpopts := cmp.Options{
-		cmpopts.IgnoreFields(Fragment{}, "ClassName"),
+		cmpopts.IgnoreFields(Fragment{}, "StyleName"),
 		cmpopts.IgnoreUnexported(Slide{}),
 	}
 
@@ -1429,7 +1429,7 @@ func TestGenerateActionsWithImages(t *testing.T) {
 	}
 
 	cmpopts := cmp.Options{
-		cmpopts.IgnoreFields(Fragment{}, "ClassName"),
+		cmpopts.IgnoreFields(Fragment{}, "StyleName"),
 		cmpopts.IgnoreUnexported(Slide{}, Image{}),
 	}
 
@@ -1624,7 +1624,7 @@ func TestAdjustSlideCount(t *testing.T) {
 
 			// Compare with expected results
 			cmpopts := cmp.Options{
-				cmpopts.IgnoreFields(Fragment{}, "ClassName"),
+				cmpopts.IgnoreFields(Fragment{}, "StyleName"),
 				cmpopts.IgnoreUnexported(Slide{}),
 			}
 
@@ -2059,7 +2059,7 @@ func TestCopySlides(t *testing.T) {
 	}
 
 	cmpopts := cmp.Options{
-		cmpopts.IgnoreFields(Fragment{}, "ClassName"),
+		cmpopts.IgnoreFields(Fragment{}, "StyleName"),
 		cmpopts.IgnoreUnexported(Slide{}),
 	}
 
@@ -2141,7 +2141,7 @@ func TestDiffSlidesDoesNotModifyOriginal(t *testing.T) {
 
 	// Check that original slides were not modified
 	cmpopts := cmp.Options{
-		cmpopts.IgnoreFields(Fragment{}, "ClassName"),
+		cmpopts.IgnoreFields(Fragment{}, "StyleName"),
 		cmpopts.IgnoreUnexported(Slide{}),
 	}
 
@@ -2778,7 +2778,7 @@ func FuzzGenerateActions(f *testing.F) {
 		}
 
 		cmpOpts := cmp.Options{
-			cmpopts.IgnoreFields(Fragment{}, "ClassName"),
+			cmpopts.IgnoreFields(Fragment{}, "StyleName"),
 			cmpopts.IgnoreUnexported(Slide{}),
 			cmpopts.EquateEmpty(),
 		}