+++ integration_action_test.go
@@ -16,7 +16,7 @@ func TestAction(t *testing.T) {
 	ctx := context.Background()
 
 	cmpopts := cmp.Options{
-		cmpopts.IgnoreFields(Fragment{}, "ClassName"),
+		cmpopts.IgnoreFields(Fragment{}, "StyleName"),
 		cmpopts.IgnoreUnexported(Slide{}),
 	}