+++ integration_apply_test.go
@@ -17,7 +17,7 @@ func TestApply(t *testing.T) {
 	ctx := context.Background()
 
 	cmpopts := cmp.Options{
-		cmpopts.IgnoreFields(Fragment{}, "ClassName"),
+		cmpopts.IgnoreFields(Fragment{}, "StyleName"),
 		cmpopts.IgnoreUnexported(Slide{}),
 	}